backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Django backend for fullstack app"
authors = ["rog olo<rogerolowski@gmail.com>"]
license = "MIT"
package-mode = false  # disables package mode for apps
# Disable package mode since this is an app, not a library
packages = []

[tool.poetry.dependencies]
django-axes = "^6.4.0"
python = "^3.10"
Django = ">=4.2,<5.0"
djangorestframework = "*"
django-extensions = "*"
django-filter = "*"
django-tailwind = "*"
django-browser-reload = "*"
django-guardian = "*"
django-braces = "*"
django-role-permissions = "*"
bleach = "*"
django-csp = "*"
django-secure = "*"
django-ipware = "*"
django-user-agents = "*"
django-extra-fields = "*"
monero = "*"
python-decouple = "*"
psycopg2-binary = "*"
redis = "*"
pillow = "*"
django-cors-headers = "*"
djangorestframework-simplejwt = "*"
drf-spectacular = "*"
django-redis = "*"
whitenoise = "*"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"
black = "*"
flake8 = "*"
isort = "*"
coverage = "*"
django-coverage-plugin = "*"
factory_boy = "*"
Faker = "*"
pytest = "*"
pytest-django = "*"
pytest-cov = "*"
django-fixture-magic = "*"
ipython = "*"
pytest-mock = "*"
pytest-env = "*"
pytest-xdist = "*"
pytest-sugar = "*"
pytest-randomly = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
